Rename test package from main to test
