+++ resources/test/grpc.go
@@ -1,4 +1,4 @@
-package main
+package test
 
 import (
 	"context"
