+++ resources/test/grpc_web.go
@@ -1,4 +1,4 @@
-package main
+package test
 
 import (
 	"bytes"
