+++ resources/test/harness.go
@@ -1,4 +1,4 @@
-package main
+package test
 
 import (
 	"encoding/json"
