+++ resources/test/utility.go
@@ -1,4 +1,4 @@
-package main
+package test
 
 import (
 	"net"
